+++ ui/v2/packages/evidently-ui/package.json
@@ -9,6 +9,12 @@
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0"
   },
   "exports": {
+    "./routes-components/*/data": {
+      "default": "./src/routes-components/*/data.tsx"
+    },
+    "./routes-components/*": {
+      "default": "./src/routes-components/*/index.tsx"
+    },
     "./*": {
       "default": "./src/*",
       "types": [
@@ -26,26 +32,35 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
-  "devDependencies": {
+  "dependencies": {
+    "tiny-invariant": "^1.3.1",
+    "react-hook-form": "^7.47.0",
+    "@hookform/resolvers": "^3.3.1",
+    "zod": "^3.22.4",
+    "dayjs": "^1.11.10",
+    "react-router-dom": "^6.16.0",
     "@nivo/bar": "^0.62.0",
     "@nivo/line": "^0.62.0",
     "@nivo/scatterplot": "^0.62.0",
     "@nivo/core": "^0.62.0",
     "@material-table/core": "^6.2.11",
     "plotly.js-cartesian-dist-min": "^2.26.2",
-    "@types/plotly.js-cartesian-dist-min": "npm:@types/plotly.js-dist-min@^2.3.2",
     "plotly.js": "^2.26.2",
+    "react-plotly.js": "^2.6.0",
+    "react-markdown": "^9.0.0",
+    "@mui/x-date-pickers": "^6.16.2"
+  },
+  "devDependencies": {
+    "@types/plotly.js-cartesian-dist-min": "npm:@types/plotly.js-dist-min@^2.3.2",
     "@types/plotly.js": "^2.12.27",
     "@types/react": "^18.2.0",
     "@types/react-dom": "^18.2.0",
-    "react-plotly.js": "^2.6.0",
     "@types/react-plotly.js": "^2.6.1",
     "@typescript-eslint/eslint-plugin": "^6.0.0",
     "@typescript-eslint/parser": "^6.0.0",
     "eslint": "^8.45.0",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.3",
-    "react-markdown": "^9.0.0",
     "typescript": "^5.0.2"
   }
 }
