+++ ui/v2/packages/evidently-ui/src/components/NavigationProgress.tsx
