+++ ui/v2/packages/evidently-ui/src/routes-components/dashboard/data.tsx
@@ -1,0 +1,35 @@
+import dayjs from 'dayjs'
+import invariant from 'tiny-invariant'
+import { DashboardInfo } from '~/api'
+import { InJectAPI } from '~/utils'
+
+export type loaderData = DashboardInfo
+
+type QueryAliases = 'FROM' | 'TO'
+export type QueryLiterals = 'date_from' | 'date_to'
+
+export const QUERY_PARAMS: Record<QueryAliases, QueryLiterals> = {
+  FROM: 'date_from',
+  TO: 'date_to'
+}
+
+export const injectAPI: InJectAPI<loaderData> = ({ api }) => ({
+  loader: ({ params, request }) => {
+    invariant(params.projectId, 'missing projectId')
+
+    const { searchParams } = new URL(request.url)
+
+    let date_from = searchParams.get(QUERY_PARAMS.FROM)
+    let date_to = searchParams.get(QUERY_PARAMS.TO)
+
+    if (date_from && !dayjs(date_from).isValid()) {
+      date_from = null
+    }
+
+    if (date_to && !dayjs(date_to).isValid()) {
+      date_to = null
+    }
+
+    return api.getProjectDashboard(params.projectId, date_from, date_to)
+  }
+})
