+++ ui/v2/packages/evidently-ui/src/routes-components/dashboard/index.tsx
@@ -1,39 +1,13 @@
 import { Alert, AlertTitle, Box, Collapse, Grid, Typography } from '@mui/material'
 import { DateTimePicker } from '@mui/x-date-pickers/DateTimePicker'
-import dayjs from 'dayjs'
-import { LoaderFunctionArgs, useLoaderData, useParams } from 'react-router-dom'
-import { DashboardContent } from 'evidently-ui/components/DashboardContent'
-import { api } from 'api/RemoteApi'
+import { useLoaderData, useParams } from 'react-router-dom'
+import { DashboardContent } from '~/components/DashboardContent'
 import invariant from 'tiny-invariant'
 import { useSearchParams } from 'react-router-dom'
-import { formatDate } from 'Utils/Datetime'
+import { formatDate } from '~/utils'
+import { QUERY_PARAMS, loaderData, QueryLiterals } from './data'
 
-type QueryAliases = 'FROM' | 'TO'
-type QueryLiterals = 'date_from' | 'date_to'
-
-const queryParams: Record<QueryAliases, QueryLiterals> = {
-  FROM: 'date_from',
-  TO: 'date_to'
-}
-
-export const loader = async ({ params, request }: LoaderFunctionArgs) => {
-  invariant(params.projectId, 'missing projectId')
-
-  const { searchParams } = new URL(request.url)
-
-  let date_from = searchParams.get(queryParams.FROM)
-  let date_to = searchParams.get(queryParams.TO)
-
-  if (date_from && !dayjs(date_from).isValid()) {
-    date_from = null
-  }
-
-  if (date_to && !dayjs(date_to).isValid()) {
-    date_to = null
-  }
-
-  return api.getProjectDashboard(params.projectId, date_from, date_to, request.signal)
-}
+import dayjs from 'dayjs'
 
 export const Component = () => {
   const { projectId } = useParams()
@@ -56,11 +30,11 @@ export const Component = () => {
     )
   }
 
-  const data = useLoaderData() as Awaited<ReturnType<typeof loader>>
+  const data = useLoaderData() as loaderData
   const dataRanges = { minDate: dayjs(data.min_timestamp), maxDate: dayjs(data.max_timestamp) }
 
-  const date_from = dayjs(searchParams.get(queryParams.FROM) || dataRanges.minDate)
-  const date_to = dayjs(searchParams.get(queryParams.TO) || dataRanges.maxDate)
+  const date_from = dayjs(searchParams.get(QUERY_PARAMS.FROM) || dataRanges.minDate)
+  const date_to = dayjs(searchParams.get(QUERY_PARAMS.TO) || dataRanges.maxDate)
 
   const isCorrectTimeInterval =
     date_from.isValid() && date_to.isValid() && date_from.isBefore(date_to)
@@ -89,7 +63,7 @@ export const Component = () => {
             }}
             label="From"
             value={date_from}
-            onChange={getOnChangeDate(queryParams.FROM)}
+            onChange={getOnChangeDate(QUERY_PARAMS.FROM)}
           />
         </Grid>
         <Grid item>
@@ -108,7 +82,7 @@ export const Component = () => {
             }}
             label="To"
             value={date_to}
-            onChange={getOnChangeDate(queryParams.TO)}
+            onChange={getOnChangeDate(QUERY_PARAMS.TO)}
           />
         </Grid>
         <Grid item xs={12}>
