+++ ui/v2/packages/evidently-ui/src/routes-components/home/data.tsx
@@ -1,0 +1,8 @@
+import { VersionInfo } from '~/api'
+import { InJectAPI } from '~/utils'
+
+export type loaderData = VersionInfo
+
+export const injectAPI: InJectAPI<loaderData> = ({ api }) => ({
+  loader: () => api.getVersion()
+})
