+++ ui/v2/packages/evidently-ui/src/routes-components/projectId/data.tsx
@@ -1,0 +1,14 @@
+import invariant from 'tiny-invariant'
+import { ProjectDetails } from '~/api'
+import { InJectAPI } from '~/utils'
+
+export type loaderData = ProjectDetails
+
+export const injectAPI: InJectAPI<loaderData> = ({ api }) => ({
+  loader: ({ params }) => {
+    const { projectId } = params
+    invariant(projectId, 'missing projectId')
+
+    return api.getProjectInfo(projectId)
+  }
+})
