+++ ui/v2/packages/evidently-ui/src/routes-components/reportId/data.tsx
@@ -1,0 +1,16 @@
+import invariant from 'tiny-invariant'
+import { DashboardInfo } from '~/api'
+import { InJectAPI } from '~/utils'
+
+export type loaderData = DashboardInfo
+
+export const injectAPI: InJectAPI<loaderData> = ({ api }) => ({
+  loader: ({ params }) => {
+    const { projectId, reportId } = params
+
+    invariant(projectId, 'missing projectId')
+    invariant(reportId, 'missing reportId')
+
+    return api.getDashboard(projectId, reportId)
+  }
+})
