+++ ui/v2/packages/evidently-ui/src/routes-components/reportId/index.tsx
@@ -1,28 +1,16 @@
-import { LoaderFunctionArgs, useLoaderData, useParams } from 'react-router-dom'
+import { useLoaderData, useParams } from 'react-router-dom'
 import invariant from 'tiny-invariant'
-import { api } from 'api/RemoteApi'
-import { DashboardContent } from 'evidently-ui/components/DashboardContent'
-import DashboardContext, {
-  CreateDashboardContextState
-} from 'evidently-ui/contexts/DashboardContext'
-import type { crumbFunction } from 'Components/BreadCrumbs'
-
-export const loader = async ({ params }: LoaderFunctionArgs) => {
-  const { projectId, reportId } = params
-
-  invariant(projectId, 'missing projectId')
-  invariant(reportId, 'missing reportId')
-
-  return api.getDashboard(projectId, reportId)
-}
-
-type loaderData = Awaited<ReturnType<typeof loader>>
+import { DashboardContent } from '~/components/DashboardContent'
+import DashboardContext, { CreateDashboardContextState } from '~/contexts/DashboardContext'
+import type { crumbFunction } from '~/components/BreadCrumbs'
+import { loaderData } from './data'
+import type { Api } from '~/api'
 
 export const handle: { crumb: crumbFunction<loaderData> } = {
   crumb: (_, { pathname, params }) => ({ to: pathname, linkText: String(params.reportId) })
 }
 
-export const Component = () => {
+export const ReportTemplate = ({ api }: { api: Api }) => {
   const { projectId, reportId } = useParams()
   invariant(projectId, 'missing projectId')
   invariant(reportId, 'missing reportId')
