+++ ui/v2/service/package.json
@@ -10,20 +10,13 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "evidently-ui": "workspace:*",
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
-    "@hookform/resolvers": "^3.3.1",
-    "@mui/icons-material": "^5.14.12",
     "@mui/material": "^5.14.12",
-    "@mui/x-date-pickers": "^6.16.2",
-    "dayjs": "^1.11.10",
-    "evidently-ui": "workspace:*",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-hook-form": "^7.47.0",
-    "react-router-dom": "^6.16.0",
-    "tiny-invariant": "^1.3.1",
-    "zod": "^3.22.4"
+    "react-router-dom": "^6.16.0"
   },
   "devDependencies": {
     "@types/react": "^18.2.0",
