+++ ui/v2/service/src/Routes/home/index.tsx
@@ -1,6 +1,22 @@
 import { RouteObject } from 'react-router-dom'
+import { injectAPI } from 'evidently-ui/routes-components/home/data'
+import { api } from 'api/RemoteApi'
+
+import { theme } from 'evidently-ui/theme/v1'
+import logoSrc from 'assets/logo.png'
+
+const { loader } = injectAPI({ api })
 
 export default {
   path: '/',
-  lazy: () => import('./Component')
+  lazy: async () => {
+    const { HomeComponentTemplate, ...rest } = await import('evidently-ui/routes-components/home')
+
+    const Component = () => {
+      return <HomeComponentTemplate theme={theme} logoSrc={logoSrc} />
+    }
+
+    return { Component, ...rest }
+  },
+  loader
 } satisfies RouteObject
