+++ ui/v2/service/src/Routes/projects/projectId/index.tsx
@@ -1,4 +1,8 @@
 import { RouteObject } from 'react-router-dom'
+import { injectAPI } from 'evidently-ui/routes-components/projectId/data'
+import { api } from 'api/RemoteApi'
+
+const { loader } = injectAPI({ api })
 
 ////////////////////
 // children routes
@@ -11,6 +15,7 @@ import TestSuitesOldRoute from './test_suites'
 
 export default {
   path: 'projects/:projectId',
-  lazy: () => import('./Component'),
+  lazy: () => import('evidently-ui/routes-components/projectId'),
+  loader,
   children: [DashboardRoute, ReportsRoute, TestSuitesRoute, TestSuitesOldRoute]
 } satisfies RouteObject
