+++ ui/v2/service/src/Routes/projects/projectId/reports/index.tsx
@@ -1,4 +1,8 @@
 import { RouteObject } from 'react-router-dom'
+import { injectAPI } from 'evidently-ui/routes-components/reports/data'
+import { api } from 'api/RemoteApi'
+
+const { loader } = injectAPI({ api })
 
 ////////////////////
 // children routes
@@ -9,6 +13,7 @@ import ReportRoute from './reportId'
 export default {
   id: 'reports',
   path: 'reports',
-  lazy: () => import('./Component'),
+  lazy: () => import('evidently-ui/routes-components/reports'),
+  loader,
   children: [ReportRoute]
 } satisfies RouteObject
