+++ ui/v2/service/src/Routes/projects/projectId/reports/reportId/index.tsx
@@ -1,7 +1,20 @@
 import { RouteObject } from 'react-router-dom'
+import { injectAPI } from 'evidently-ui/routes-components/reportId/data'
+import { api } from 'api/RemoteApi'
+
+const { loader } = injectAPI({ api })
 
 export default {
   id: 'show-report-by-id',
   path: ':reportId',
-  lazy: () => import('./Component')
+  lazy: async () => {
+    const { ReportTemplate, ...rest } = await import('evidently-ui/routes-components/reportId')
+
+    const Component = () => {
+      return <ReportTemplate api={api} />
+    }
+
+    return { Component, ...rest }
+  },
+  loader
 } satisfies RouteObject
