+++ ui/v2/service/src/Routes/projects/projectId/test-suites/index.tsx
@@ -1,10 +1,15 @@
 import { RouteObject } from 'react-router-dom'
+import { injectAPI } from 'evidently-ui/routes-components/test-suites/data'
+import { api } from 'api/RemoteApi'
 
 import TestSuiteRoute from './testSuiteId'
+
+const { loader } = injectAPI({ api })
 
 export default {
   id: 'test_suites',
   path: 'test-suites',
-  lazy: () => import('./Component'),
+  lazy: () => import('evidently-ui/routes-components/test-suites'),
+  loader,
   children: [TestSuiteRoute]
 } satisfies RouteObject
