+++ ui/v2/service/src/Routes/projectsList/index.tsx
@@ -1,6 +1,12 @@
 import { RouteObject } from 'react-router-dom'
+import { injectAPI } from 'evidently-ui/routes-components/projectsList/data'
+import { api } from 'api/RemoteApi'
+
+const { loader, action } = injectAPI({ api })
 
 export default {
   index: true,
-  lazy: () => import('./Component')
+  lazy: () => import('evidently-ui/routes-components/projectsList'),
+  loader,
+  action
 } satisfies RouteObject
