+++ ui/v2/service/src/main.tsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom/client'
 import './index.css'
 
 import { createBrowserRouter, RouterProvider } from 'react-router-dom'
-import { Routes } from 'Routes'
+import { Routes } from 'routes'
 
 const router = createBrowserRouter(Routes)
